+++ ChangeLog.md
@@ -22,6 +22,9 @@ Other enhancements:
   `extra-deps` of `stack.yaml`
 * `stack build` suggests trying another GHC version should the build
   plan end up requiring unattainable `base` version.
+* A new sub command `run` has been introduced to build and run a specified executable
+  similar to `cabal run`. If no executable is provided as the first argument, it
+  defaults to the first available executable in the project.
 * `stack build` missing dependency suggestions (on failure to construct a valid
   build plan because of missing deps) are now printed with their latest
   cabal file revision hash. See